Ethical and legal challenges of the darkweb
As a part of an ongoing discussion, I addressed the challenges of the "darkweb" at WSIS 2017. A high-level discussion followed with public policy professionals and cybercrime researchers as to how relevant is the idea of regulating this portion of the Internet. Link to the session: https://www.itu.int/net4/wsis/forum/2017/Agenda/Session/257#intro

File is also uploaded at Slideshare https://www.slideshare.net/MaliciaRogue/ethical-and-legal-challenges-of-the-darkweb